Price decimal to int

+++ App/Application/DTOS/Product/ProductResponse.cs
@@ -9,5 +9,5 @@ public class ProductResponseDTO
     public required string Name { get; set; }
 
     [Required]
-    public required decimal Price { get; set; }
+    public required int Price { get; set; }
 }
